DT 4.0 might use fixed headroom instead of _opencl_get_unused_device_mem() for tuning memory
As a reminder for those not fluent in cl memory management.

The pipeline code needs to know how much cl memory might be used by a module to decide about
1. no need to tile or 2. have to tile and how large the tiles can be, this is basically about avoiding
cpu fallbacks and thus performance drops.

The memory that dt is allowed to take is available memory (AM), this used to be calculated as
  physical memory (PM) - headroom (was a conf value)

Now we calculate it in non-tuned modes as a fraction of PM

If we switch on "tuning memory" we look for a conf key and
a) if it is a positive int we take that as a fixed headroom per device, so almost the same as in pre 4.0
b) we check for available memory on the card by some sort of "brute force" check in
     _opencl_get_unused_device_mem()
   so users don't have to know or tell dt what memory other apps or the OS use.

So what is the problem?
At least for nvidia 515/16 drivers we could identify a problem exactly in the code for b)
The reason for this is not understood yet so for a stable release we opt for a fixed headroom
as default for new users or thos updating dt.