Be more optimistic for too slow OpenCL cards.
As hinted by @kofa73 there has been for long a check for very slow OpenCL
devices and if so OpenCL would be switched off by default if darktable is
updated to 4.0 or after a fresh install.

With current CPUs benchmark performance the ratio used for disabling OpenCL
seems to be wrong.

Here we switch if off if the cpu is 3 times faster than GPU.