Shortcuts : handle key modifiers and numpad properly
On AZERTY and BÉPO keyboards, numbers are accessed by shift + &, shift + é, shift + ", etc. They should be referenced as the system sees them, not as they are laid out on the keyboard. Then, the modifiers should be removed from the shortcut.

Adds numpad alternatives for default rating shortcuts because WHYYYYYY ?