Filmic RGB : fix overzealous clipping
v3 to 5: the output of the filmic spline was clipped in [0;1]. That's not HDR-able (aka future-proof) and also not entirely safe for black point compensation (in case the spline under-shoot).

v6 and 7: the output of the filmic spline was clipped to display black/white __before__ applying power, but using values corrected for power. Given that the power needs at least a clipping removing negatives, clip to the min/max of the spline instead.
This bug showed when setting display white < 100% to print on Fuji Instax and resulted in hash clipping with flat output.

None of these bugs affected people outputting RGB to [0; 100] %, that is the default settings.