Image.c : reintroduce local copy fetching in dt_image_write_sidecar_file
Not having it seems to make XMP writing fail randomly when importing new pictures.
That makes no sense considering the original file is still being fetched and should be present at import time.
We are most likely looking at a deeper problem buried in the depths of the function tree.