Bauhaus: Fix wrongly floored value of percentage sliders
When format of a slider is set to percentage, it will decrease 2 digits
and set factor to 100, so when we set normalized value of the slider,
we need to handle both digits and factor.

In the large refator of commit 48a86a9c0353f87444fd6956f087af933ecad599,
we forget to multiply factor while setting normalized value, this makes
percentage sliders always get floored to integer value and lose the
float value.

This commit fixes it by multiplying factor while setting normalized
value.