Bauhaus : fix a bug in coordinates handling for popups
Gtk is weird, it sends motion_notify events on widgets that are not hovered.
Since popups are layed over, they behave differently and we can't use their relative coordinates, because there is a chance they belong to another widget.
For popups, we grab absolute coordinates in main window space instead.