pipeline cache : simplify the code, fix inconsistent states ?
The pipe cache code was a tangled mess taking input from GUI (active module) and pipeline. That's prone to errors. Plus it had a couple of smart-ass perks (ROI-independent cache ? what for ?).

Changes made : 
- each module gets a global hash representing its internal params, blending params, previous modules and ROI size.
- this hash is attributed at pipeline (re)construction and independant from GUI events. It gets recomputed on the whole pipeline.
- this hash is used as a key to fetch cache lines.

Any upstream change in a module will automatically be propagated to the next hashes. Therefore, we only need to find the latest valid hash (starting from the end of the pipeline) and start recomputing from there. No need to track what module the user was tingling when recomputation was requested.