Rework the pipeline vs. GUI synchronization.
Manually trigger pipeline recomputes on GUI events that require it. Ensure redrawings are dispatched before pipe recomputes for better UX.

The previous logic was to hide the pipe recompute into the darkroom `expose()` function, used as a Gtk callback for Cairo redrawing of the center view. Events that would invalidate the history would set `pipe->*_status = DIRTY` in a non-thread-safe way, through the `dt_dev_invalidate()` methods, then request a redraw of the center region, and `expose()` would listen to that and fire a pipeline at drawing time. That was absolutely useless, because the image was not ready at drawing time, so the same `expose()` function would have to be connected to the *_PIPE_FINISHED signal and redraw once again then.

It took me a full afternoon to figure out what the fuck did start those computations, and that's the last place I expected it to be, because it makes no sense to rely on GUI events to start background tasks. Namely, if the GUI froze (because of too many user interactions restarting too many pipes), then the pipe would freeze too. That's like all the trouble of multi-threading without any of the benefits.

The present logic is more verbose but makes the 3 main events independent from each others :
- commit history and invalidate pipeline if changes,
- redraw UI, especially to display the "busy..." message,
- spawn pipe recomputations.

More importantly, it allows to have multiple history invalidations at the same time, while only triggering pipeline once, or re-launching it with the kill-switch. Then, the GUI grabs the image backbufer at its own pace, with signals. This makes the whole process more predictable and voids all the nasty workarounds piled up over the years to flush the cache all the time, because we were unable to assert its validity.