CMake: Clear RPATH mess
Use PLUGINDIR for $LIBDIR/ansel, because this is the name used in
function name, and then we could use LIBDIR for the system lib dir.

Install libansel.so to LIBDIR, this allows us to totally disable RPATH
with CMAKE_SKIP_RPATH when it will be installed as system software
because we only have other libraries linked to it.

Correct RPATH for all binaries and libraries, because $ORIGIN is current
libraries dir.