CMake: Drop run-from-build-dir
In order to support run-from-build-dir[0], what basically did is
"manually do what will be done during installation in the build time",
that contains a mixture of change binary_dir of "change binary_dir of
subdirectories to pretend it as the same structure of install dir" and
"manually copy files to build dir to pretend it as the same structure of
install dir". This changes the structure of build dir, adds a lot of
variables into CMake files, and often use variables wrongly.

Because this makes CMake files hard to read (the variables are
misleading), and running desktop apps directly from build dir is not a
recommended way to user, this commit will drop it.

User should always install it before running it, it is easy to install
to where a normal user has permission by changing
`CMAKE_INSTALL_PREFIX`.

Loading resources and plugins relatively is kept so the app is still
relocatable.

[0]: https://github.com/darktable-org/darktable/pull/5086