dev: add history item : don't play smart-ass on history collapsing
Don't try to detect if we should merge the current history item with the one before, as it triggers all sorts of edge cases with drawn masks.

Every history step is now recorded, which might pollute the history with intermediate steps, but the purpose of an history is not to decide what should be kept or merged. Merging happens with the GUI button.