Bauhaus: rework the timeout logic
Commit history after 350 ms no matter what. If another user interaction happens in those 350 ms, destroy the previous timer and start again. That leaves 350 ms to user to complete another scrolling step.

May not be enough for senior citizens, and too much for gamers, in which case this needs to be an user config.