GUI pipelines : force them to wait for each other
Allowing GUI pipelines to run concurrently would be nice with several GPUs. Problem is not all modules use OpenCL, so both pipes would still have to share CPU, which is a bottleneck.

Forcing pipelines to run in sequence and to wait for each other not only reduces the individual wall-clock runtime, but also the cumulative one, meaning users will get a full GUI refresh (much) sooner, especially on weak hardware.

That may also allow to (re)build the pipeline nodes just once for both pipes in the future.