Ashift: make straightening mode and editing mode mutually exclusive
Straightening mode is the ported thing from clip.c module, where horizon line is drawn from holding right click on the image.

Editing is the native mode from this module, where we work on content lines.

Straightening mode is somewhat of a duplicate feature that collides with the native mode because both modes use the right click button_pressed event with a different meaning. Bypass straightening when full editing mode is on.