Fix #352
Not sure what shitty side-effect of Bauhaus comboboxes was used to ensure history was committed when setting the filepath to the combobox text… Side-effect was likely removed, so of course it broke lazy-ass code.

The reason why we mix comboboxes with native filechooser here is beyond me. It's piling the drawbacks of both methods.