dev history_end: clamp at 0 instead of 1
Clamping at 0 means no history at all but only raw image. This is not supposed to be a valid value, because they should be starting at 1. Except that if we clamp at 1, there is an offset in GUI. No time to troubleshoot the twisted logic of what's going on here.