History: revert to collapsing history steps if the module didn't change
This sucks: it means that history steps don't match the undo/redo list of actions. But the current way of adding another history step for each user interaction made the history grow out of proportions, especially when using masks, and that ends up saved in database + XMP.

The real problem here is that the masks history is included into the module history, so we have an history length of (n module states × m masks states). The real solution would be to maintain a mask history separately, so we store an history of (n module states) and another of (m masks states), that is a total of (n + m) instead of (n × m) entries, and then linking masks and modules states by ID (relational database 101).

But that requires invalidating the database scheme, so it's for a future major version.