dt_dev_read_history: don't write...
dt_dev_read_history also sanitizes IOP modules order and updates blendop and module params structures to newer versions if needed. So, writing back the result of the sanitization might make sense to avoid re-doing it the next time.

But there are many reasons for which we don't want to:
- it's twice as many I/O, when opening a new picture, and that hurts for images stored on NAS and histories with lots of masks,
- history is saved every 30s in background thread while the darkroom is opened, and when the darkroom is closed or changes active picture,
- reading history is done to refresh thumbnails or to batch export, which may be done from a future version of the software, that will therefore make the histories incompatible with older versions. That's not transparent to the user and not expected, as long as user didn't edit anything.
- the function is called "read", goddamit, can we please stick to that ?