Darkroom: prevent useless recomputations of main preview when entering the view
Darkrom `configure()` method is connected to `"configure-event"` event, which is called when initing the view AND upon window resizes events (through Gtk widget/window resize commands).

Problem is, at view init time, the final window size may not be correct just yet. It will be when we call `dt_ui_restore_panels()`, which will resize stuff properly.  That only happens when entering the current view.

Until we run `dt_dev_configure()`, main preview pipe gets output size -1×-1 px, which aborts the pipe recompute early. As soon as we init sizes with something "valid" with regard to the pipe, pipeline runs. Problem is it will not be valid with regard to the (final) window size and the output of the pipe will be thrown out, until we get the final size.

TD;DR: until we get the final window size, which happens only when entering the view, don't configure the main preview pipeline, which will disable useless recomputes.