Resync history top: handle cases where there is more than one history items on top to resync
The sync_top method is the fast track for when we know regular stuff happened (user added history entries by performing changes in sliders/comboboxes, etc.). It doesn't apply for history merging or reverting to previous history steps. In that scenario, the current code expected that only one history item would need resync, but on slow systems, it is possible that users have added more history entries during a pipe recompute. We handle this case here by using the history hash, record the hash of the last-known history, and then update everything from top to this history step until we find its hash on the stack.